extracted more functions into separate modules
