+++ app.js
@@ -1,6 +1,7 @@
-const fs = require("fs");
 const rimraf = require("rimraf");
-const puppeteer = require("puppeteer");
+
+const puppeteerConnect = require("./src/puppeteer_connect");
+const createDirectories = require("./src/create_directories");
 const screenshot = require("./src/screenshot");
 const saveReport = require("./src/save_report");
 
@@ -13,56 +14,14 @@ if (!url) {
 	url = "http://localhost:3000";
 }
 
-const puppeteerConnect = async (url) => {
-	const browser = await puppeteer.launch({
-		// headless: false,
-		// slowMo: 250,
-	});
+const runApp = async (url) => {
+	const { browser, page } = await puppeteerConnect.puppeteerConnect(url);
 
-	const page = await browser.newPage();
-	await page.setViewport({ width: 1980, height: 1000 });
-	page.setDefaultTimeout(10000);
-
-	try {
-		await page.goto(url, {
-			waitUntil: "networkidle2",
-		});
-	} catch (err) {
-		console.log(
-			"Having trouple establishing a connection. If you are attempting to connect to localhost try entering the url as your local ip address"
-		);
-		console.log(err);
-		browser.close();
-		return;
-	}
-
-	// helper function to wait for a specified period of time
-	const delay = (time) => {
-		return new Promise(function (resolve) {
-			setTimeout(resolve, time);
-		});
-	};
-
-	// get the current date in YYYY-MM-DD format (for creating directories)
-	let timeStamp = Date.now();
-	let dateObject = new Date(timeStamp);
-	let date = dateObject.getDate();
-	let month = dateObject.getMonth() + 1;
-	let year = dateObject.getFullYear();
-
-	const imagesDirectory = `./outputs/${
-		year + "-" + month + "-" + date + "-" + timeStamp
-	}`;
-	const reportsDirectory = `${imagesDirectory}/reports`;
-
-	if (!fs.existsSync(reportsDirectory)) {
-		fs.mkdirSync(reportsDirectory, { recursive: true });
-	}
-
-	const tempImagesDirectory = "./outputs/tempImages";
-	if (!fs.existsSync(tempImagesDirectory)) {
-		fs.mkdirSync(tempImagesDirectory, { recursive: true });
-	}
+	const {
+		imagesDirectory,
+		reportsDirectory,
+		tempImagesDirectory,
+	} = await createDirectories.createDirectories();
 
 	// take screenshots of charts
 	try {
@@ -98,4 +57,4 @@ const puppeteerConnect = async (url) => {
 	browser.close();
 };
 
-puppeteerConnect(url);
+runApp(url);
